feat: Provide specialised adapters for both Next.js routers
The unified adapter is still available, in case it gets mounted in a component
that needs to be router-agnostic (can be mounted in either router).
Using the specialised adapters shaves around 400B off shared chunks.
