+++ packages/nuqs/src/adapters/next/app.ts
@@ -1,0 +1,4 @@
+import { createAdapterProvider } from '../internal.context'
+import { useNuqsNextAppRouterAdapter } from './impl.app'
+
+export const NuqsAdapter = createAdapterProvider(useNuqsNextAppRouterAdapter)
