+++ packages/nuqs/src/adapters/next/shared.ts
@@ -1,0 +1,8 @@
+import { renderQueryString } from '../../url-encoding'
+
+export function renderURL(base: string, search: URLSearchParams) {
+  const hashlessBase = base.split('#')[0] ?? ''
+  const query = renderQueryString(search)
+  const hash = location.hash
+  return hashlessBase + query + hash
+}
